+++ ak_library_management_orm/__manifest__.py
@@ -2,7 +2,7 @@
 
 {
     'name': 'library management ORM',
-    'version': '19.0.1.0.1',
+    'version': '19.0.1.1.0',
     'summary': """Manage Books""",
     'description': """We can Perform ORM method operations.""",
     'category': 'Library management',