+++ ak_library_management_orm/changelog.md
@@ -9,3 +9,8 @@ All notable changes to this project will be documented in this file.
 - Updated all copy methods.
 - Updated the UserError string and logic.
 - Removed unnecessary code and comment lines.
+
+[19.0.1.1.0] :
+- Added two sequence 1)library.book, 2)author.author.
+- Updated author form view: set new field in left-top with h1 teg.
+- Updated book form view: make isbn field to readonly.